# [sql_lab] Improve performance, only use slow func when needed